eon v4.0.0 Eon

Eon is a small library for using .exs files as a document store. Files read with Eon are expected to contain only an Elixir map as well as execute no arbitary code unless specified to do so. Eon can also write maps and structs to file.

Eon is useful for when you would normally store data as JSON but you need to preserve Elixir’s datatypes. Functions, ports and other datatypes that cannot be represented as pure data are not supported by Eon.

Functions

read(filename)

Loads a file, expecting a file with a single map that executes no arbitrary code. Returns a tuple which contains :ok or :error as the first element and the result or error message respectively.

Examples

iex> Eon.read("test/fixtures/basic.exs")
{:ok, %{hello: "world"}}

iex> Eon.read("test/fixtures/unsafe.exs")
{:error, :unsafe}

iex> Eon.read("non_existent_file.exs")
{:error, :enoent}

read!(filename)

Same as read/1 except it returns the result rather than an :ok tuple containing it. In the case of an error, an exception is raised.

Examples

iex> Eon.read!("test/fixtures/basic.exs")
%{hello: "world"}

iex> Eon.read!("test/fixtures/unsafe.exs")
** (Eon.Error) File results in code execution. Use Eon.read_unsafe! to bypass.

iex> Eon.read!("non_existent_file.exs")
** (Eon.Error) No such file or directory.

read_unsafe(filename)

Same as read/1 except allows arbitrary code execution. This effectively bypasses the step which prohibits the execution of Elixir code containing potentially unsafe data structures.

read_unsafe(filename, bindings)

Same as read_unsafe/1 except it takes a map as a second argument. Unbound variables within the loaded file that match a key within the passed map will be replaced with the corresponding value.

read_unsafe!(filename)

Same as read_unsafe/1 except it returns the result rather than an :ok tuple containing it. In the case of an error, an exception is raised.

read_unsafe!(filename, bindings)

Same as read_unsafe/2 except it returns the result rather than an :ok tuple containing it. In the case of an error, an exception is raised.

write(map, filename)

Takes a map and writes it to file, returning the port of the IO connection.

Examples

iex> Eon.write(%{hello: "world"}, "hello.exs")
{:ok, #PID<0.159.0>}

iex> Eon.write(%{hello: "world"}, "/hello.exs")
{:error, :eacces}

write!(map, filename)

Same as write/2 except it returns the result rather than an :ok tuple containing it. In the case of an error, an exception is raised.

Examples

iex> Eon.write(%{hello: "world"}, "hello.exs")
#PID<0.159.0>

iex> Eon.write(%{hello: "world"}, "/hello.exs")
** (Eon.Error) Permission denied.